# api/internal/runtime/static/openapi.yaml
openapi: 3.0.3
info:
  title: NAAP Analytics API
  version: "1"
  description: |
    Read-only analytics API for the Livepeer AI network.
    Backed by ClickHouse materialized views fed from two Kafka topics:
    `network_events` (daydream org) and `streaming_events` (cloudspe org).

    **Rate limiting:** 100 req/s sustained, burst 200 per IP.
    Exceeded requests receive `429 Too Many Requests`.

    **Default time window:** last 24 hours when `start`/`end` are omitted.

    **WEI values** (payment amounts) are returned as decimal strings to avoid
    float64 precision loss (e.g. `"1234500000000000000"`).
servers:
  - url: https://naap-api.cloudspe.com
    description: Production
  - url: http://localhost:8000
    description: Local development
  - url: https://naap-api.livepeer.cloud
    description: Production
...
components:
  parameters:
    org:
      name: org
      in: query
      description: Filter to a single organisation. Omit for all orgs.
      schema:
        type: string
        enum: [daydream, cloudspe]
    start:
      name: start
      in: query
      description: Window start (RFC3339). Defaults to 24 hours ago.
      schema:
        type: string
        format: date-time
        example: "2026-03-01T00:00:00Z"
    end:
      name: end
      in: query
      description: Window end (RFC3339). Defaults to now.
      schema:
        type: string
        format: date-time
        example: "2026-03-24T23:59:59Z"
    limit:
      name: limit
      in: query
      description: Maximum rows to return (1–1000).
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
    offset:
      name: offset
      in: query
      description: Pagination offset.
      schema:
        type: integer
        minimum: 0
        default: 0
    pipeline:
      name: pipeline
      in: query
      description: Filter by pipeline name (e.g. `text-to-image`).
      schema:
        type: string
    orch_address:
      name: orch_address
      in: query
      description: Filter by orchestrator Ethereum address (lowercased).
      schema:
        type: string
        example: "0xabc123def456"
...
    stream_id:
      name: stream_id
      in: query
      schema:
        type: string
...
    address_path:
      name: address
      in: path
      required: true
      description: Orchestrator Ethereum address (lowercased).
      schema:
        type: string
        example: "0xabc123def456"
  schemas:
...
    Problem:
      description: RFC 7807 Problem Details error response.
      type: object
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Too Many Requests"
        status:
          type: integer
          example: 429
        detail:
          type: string
          example: "rate limit exceeded"
    WEI:
      description: |
        Payment value in wei (1e-18 ETH), returned as a decimal string to
        preserve full uint64 precision.
      type: string
      pattern: "^[0-9]+$"
      example: "1234500000000000000"
...
    OrgPaymentTotal:
      type: object
      properties:
        TotalWEI:
          $ref: "#/components/schemas/WEI"
        Count:
          type: integer
          format: int64
    PaymentSummary:
      type: object
      properties:
        StartTime:
          type: string
          format: date-time
        EndTime:
          type: string
          format: date-time
        TotalPaymentsWEI:
          $ref: "#/components/schemas/WEI"
        PaymentEventCount:
          type: integer
          format: int64
        UniqueOrchestrators:
          type: integer
          format: int64
        ByOrg:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/OrgPaymentTotal"
    PaymentBucket:
      type: object
      properties:
        Timestamp:
          type: string
          format: date-time
        TotalWEI:
          $ref: "#/components/schemas/WEI"
        EventCount:
          type: integer
          format: int64
        UniqueOrchs:
          type: integer
          format: int64
    PipelinePayment:
      type: object
      properties:
        Pipeline:
          type: string
        TotalWEI:
          $ref: "#/components/schemas/WEI"
        EventCount:
          type: integer
          format: int64
        AvgPriceWeiPerPixel:
          type: number
          format: double
    OrchPayment:
      type: object
      properties:
        Address:
          type: string
        TotalReceivedWEI:
          $ref: "#/components/schemas/WEI"
        PaymentCount:
          type: integer
          format: int64

    # ── Reliability (R5) ────────────────────────────────────────────────────
...
    Gateway:
      type: object
      properties:
        Address: { type: string }
        Name: { type: string }
        Avatar: { type: string }
        DepositWEI: { $ref: "#/components/schemas/WEI" }
        ReserveWEI: { $ref: "#/components/schemas/WEI" }
        UpdatedAt: { type: string, format: date-time }
        ActiveStreams: { type: integer, format: int64 }
    GatewayProfile:
      type: object
      properties:
        Address: { type: string }
        Name: { type: string }
        Avatar: { type: string }
        DepositWEI: { $ref: "#/components/schemas/WEI" }
        ReserveWEI: { $ref: "#/components/schemas/WEI" }
        UpdatedAt: { type: string, format: date-time }
        StreamsRouted: { type: integer, format: int64 }
        ActiveStreams: { type: integer, format: int64 }
        TotalPaymentsWEI: { $ref: "#/components/schemas/WEI" }
        OrchsUsed:
          type: array
          items: { type: string }
    GatewayOrch:
      type: object
      properties:
        OrchAddress: { type: string }
        Name: { type: string }
        StreamCount: { type: integer, format: int64 }
        LastSeen: { type: string, format: date-time }

    # ── Stream Extensions (R8) ───────────────────────────────────────────────
...
    ModelPerformance:
      type: object
      properties:
        ModelID: { type: string }
        Pipeline: { type: string }
        AvgFPS: { type: number, format: double }
        P50FPS: { type: number, format: double }
        P99FPS: { type: number, format: double }
        WarmOrchCount: { type: integer, format: int64 }
        TotalStreams: { type: integer, format: int64 }
    OrchModelStats:
      type: object
      properties:
        Address: { type: string }
        Name: { type: string }
        AvgFPS: { type: number, format: double }
        IsWarm: { type: boolean }
    ModelDetail:
      type: object
      properties:
        ModelID: { type: string }
        Pipeline: { type: string }
        AvgFPS: { type: number, format: double }
        P50FPS: { type: number, format: double }
        P99FPS: { type: number, format: double }
        WarmOrchCount: { type: integer, format: int64 }
        TotalStreams: { type: integer, format: int64 }
        Orchs:
          type: array
          items: { $ref: "#/components/schemas/OrchModelStats" }

    # ── Extended Payments (R13) ──────────────────────────────────────────────
    GatewayPayment:
      type: object
      properties:
        GatewayAddress: { type: string }
        Name: { type: string }
        TotalWEI: { $ref: "#/components/schemas/WEI" }
        EventCount: { type: integer, format: int64 }
        UniqueOrchs: { type: integer, format: int64 }
    StreamPayment:
      type: object
      properties:
        StreamID: { type: string }
        Org: { type: string }
        Pipeline: { type: string }
        TotalWEI: { $ref: "#/components/schemas/WEI" }
        EventCount: { type: integer, format: int64 }

    # ── Failure Analysis (R14) ───────────────────────────────────────────────
...
    CapacityEntry:
      type: object
      properties:
        Pipeline: { type: string }
        ModelID: { type: string }
        WarmOrchCount: { type: integer, format: int64 }
        ActiveStreams: { type: integer, format: int64 }
        UtilizationPct: { type: number, format: double, description: "active streams / warm orchs (0–1)" }
        TotalVRAMBytes: { type: integer, format: int64 }
    CapacitySummary:
      type: object
      properties:
        SnapshotTime: { type: string, format: date-time }
        Entries:
          type: array
          items: { $ref: "#/components/schemas/CapacityEntry" }
  responses:
    429:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    500:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    BadRequest:
      description: Bad request.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }

# ── Paths ──────────────────────────────────────────────────────────────────────
paths:
...
  /v1/payments/summary:
    get:
      tags: [Payments]
      summary: Payment summary
      description: Total payments in WEI, event count, and unique orchestrators over the window. Broken down by org.
      operationId: getPaymentSummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/history:
    get:
      tags: [Payments]
      summary: Payment history
      description: Hourly payment totals for charting.
      operationId: listPaymentHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/by-pipeline:
    get:
      tags: [Payments]
      summary: Payments by pipeline
      description: Total payments and average price per pixel for each pipeline.
      operationId: listPaymentsByPipeline
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelinePayment"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/by-orch:
    get:
      tags: [Payments]
      summary: Payments by orchestrator
      description: Total payments received per orchestrator, ordered by total descending.
      operationId: listPaymentsByOrch
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrchPayment"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Reliability ────────────────────────────────────────────────────────────
...
  /v1/net/gateways:
    get:
      summary: List gateways
      description: Returns all known gateways with active stream counts.
      operationId: listGateways
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Gateway" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/gateways/{address}:
    get:
      summary: Gateway profile
      description: Returns stats and metadata for one gateway.
      operationId: getGatewayProfile
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayProfile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/gateways/{address}/orchs:
    get:
      summary: Gateway orchestrators
      description: Returns orchestrators that handled traffic for a given gateway.
      operationId: listGatewayOrchestrators
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/address_path"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GatewayOrch" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Stream Extensions (R8) ───────────────────────────────────────────────
...
  /v1/perf/by-model:
    get:
      summary: Performance by AI model
      description: Returns FPS performance broken down by AI model, joined from GPU inventory.
      operationId: listModelPerformance
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModelPerformance" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/model:
    get:
      summary: Model detail
      description: Returns detail for one AI model including per-orch FPS stats. Use `model_id` query param since model IDs contain slashes.
      operationId: getModelDetail
      tags: [Network, Performance]
      parameters:
        - name: model_id
          in: query
          required: true
          description: AI model identifier (e.g. `stabilityai/stable-diffusion-xl-base-1.0`).
          schema: { type: string }
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ModelDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Extended Payments (R13) ──────────────────────────────────────────────
  /v1/payments/by-gateway:
    get:
      summary: Payments by gateway
      description: Returns payment totals aggregated by originating gateway address.
      operationId: listPaymentsByGateway
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GatewayPayment" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/payments/by-stream:
    get:
      summary: Payments by stream
      description: Returns total payments aggregated per stream session.
      operationId: listPaymentsByStream
      tags: [Payments]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/stream_id"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StreamPayment" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Failure Analysis (R14) ───────────────────────────────────────────────
...
  /v1/net/capacity:
    get:
      summary: Network capacity vs. demand
      description: Returns GPU supply (warm orchs, VRAM) vs. active stream demand per (pipeline, model).
      operationId: getCapacitySummary
      tags: [Capacity, Network]
      parameters:
        - $ref: "#/components/parameters/org"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CapacitySummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }